Memory manager is finished, bootloader reworked. Did a bit of refactoring

+++ kernel/src/arch/amd64/bit_ops.rs
@@ -1,3 +1,3 @@
 use core::mem::size_of;
 
-pub const USIZE_BITS: usize = size_of::<usize>() * 8;
+pub const USIZE_BITS: u64 = size_of::<usize>() as u64 * 8;
